refactor : 주문 datasource 추상화 및 네트워크 관련 로직 이동
