+++ app/src/main/java/woowacourse/shopping/data/order/OrderDataSource.kt
@@ -1,0 +1,10 @@
+package woowacourse.shopping.data.order
+
+import woowacourse.shopping.data.NetworkResult
+
+interface OrderDataSource {
+    fun requestOrder(
+        cartItemIds: List<Int>,
+        callBack: (NetworkResult<Unit>) -> Unit,
+    )
+}
