+++ app/src/main/java/woowacourse/shopping/data/order/RemoteOrderDataSource.kt
@@ -1,14 +1,36 @@
 package woowacourse.shopping.data.order
 
 import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
+import woowacourse.shopping.data.NetworkResult
 import woowacourse.shopping.data.dto.request.OrderRequest
 import woowacourse.shopping.data.remote.ApiClient
 
-class RemoteOrderDataSource {
+class RemoteOrderDataSource(
     private val orderApiService: OrderApiService =
-        ApiClient.getApiClient().create(OrderApiService::class.java)
+        ApiClient.getApiClient().create(OrderApiService::class.java),
+) : OrderDataSource {
+    override fun requestOrder(
+        cartItemIds: List<Int>,
+        callBack: (NetworkResult<Unit>) -> Unit,
+    ) {
+        orderApiService.requestOrder(OrderRequest(cartItemIds)).enqueue(
+            object : Callback<Unit> {
+                override fun onResponse(
+                    call: Call<Unit>,
+                    response: Response<Unit>,
+                ) {
+                    callBack(NetworkResult.Success(Unit))
+                }
 
-    fun requestOrder(cartItemIds: List<Int>): Call<Unit> {
-        return orderApiService.requestOrder(OrderRequest(cartItemIds))
+                override fun onFailure(
+                    call: Call<Unit>,
+                    t: Throwable,
+                ) {
+                    callBack(NetworkResult.Error)
+                }
+            },
+        )
     }
 }
